ObterNacionalidade finalizado

+++ src/VilevePay.Application/ViewModels/v1/Parametrizacao/NacionalidadeViewModel.cs
@@ -2,7 +2,7 @@
 {
     public class NacionalidadeViewModel
     {
-        public int CodigoNacionalidade { get; set; }
+        public long CodigoNacionalidade { get; set; }
         public string Nome { get; set; }
         public string PrefixoPais { get; set; }
     }
